wifi: Use 16 bits to represent the size of Information fields
...to support fragmentation of IEs that do not fit in 254/255 bytes

+++ src/mesh/model/dot11s/ie-dot11s-peer-management.h
@@ -115,9 +115,9 @@ public:
 
   // Inherited from WifiInformationElement
   virtual WifiInformationElementId ElementId () const;
-  virtual uint8_t GetInformationFieldSize (void) const;
+  virtual uint16_t GetInformationFieldSize (void) const;
   virtual void SerializeInformationField (Buffer::Iterator i) const;
-  virtual uint8_t DeserializeInformationField (Buffer::Iterator i, uint8_t length);
+  virtual uint16_t DeserializeInformationField (Buffer::Iterator i, uint16_t length);
   virtual void Print (std::ostream& os) const;
 
 private:
