+++ src/mesh/model/dot11s/ie-dot11s-perr.h
@@ -67,9 +67,9 @@ public:
   // Inherited from WifiInformationElement
   virtual WifiInformationElementId ElementId () const;
   virtual void SerializeInformationField (Buffer::Iterator i) const;
-  virtual uint8_t DeserializeInformationField (Buffer::Iterator start, uint8_t length);
+  virtual uint16_t DeserializeInformationField (Buffer::Iterator start, uint16_t length);
   virtual void Print (std::ostream& os) const;
-  virtual uint8_t GetInformationFieldSize () const;
+  virtual uint16_t GetInformationFieldSize () const;
 
 private:
   std::vector<HwmpProtocol::FailedDestination> m_addressUnits; ///< address units
